Rename `parsers` module to `formats`
