+++ src/lib.rs
@@ -1,9 +1,9 @@
 pub const INOCHI2D_SPEC_VERSION: &str = "1.0-alpha";
 
+pub mod formats;
 pub mod math;
 pub mod mesh;
 pub mod model;
 pub mod nodes;
-pub mod parsers;
 pub mod puppet;
 pub mod renderers;
